feat(build): Improved building of email templates
This changes how the email templates are built such that each individual
template and its corresponding txt or html output are a single CMake
target. This way the build can be more easily parallelized to speed
things up.
