+++ emails/src/emails/ForgotPassword/index.tsx
@@ -8,8 +8,8 @@ import {
   Text,
 } from '@react-email/components';
 
-import EmailLayout from '../components/EmailLayout';
-import EmailLogo from '../components/EmailLogo';
+import EmailLayout from '../../components/EmailLayout';
+import EmailLogo from '../../components/EmailLogo';
 
 interface ForgotPasswordProps {
   baseUrl?: string;
