+++ emails/src/emails/TrialAboutToExpire/index.tsx
@@ -1,57 +1,52 @@
 import * as React from 'react';
 import {
-  Button,
   Heading,
   Hr,
   Link,
-  Section,
   Text,
 } from '@react-email/components';
 
-import EmailLayout from '../components/EmailLayout';
-import EmailLogo from '../components/EmailLogo';
+import EmailLayout from '../../components/EmailLayout';
+import EmailLogo from '../../components/EmailLogo';
 
-interface PlaidDisconnectedProps {
+interface TrialAboutToExpireProps {
   baseUrl?: string;
   firstName?: string;
   lastName?: string;
-  linkName?: string;
-  linkURL?: string;
+  trialExpirationDate?: string;
+  trialExpirationWindow?: string;
   supportEmail?: string;
 }
 
-export const PlaidDisconnected = ({
+export const TrialAboutToExpire = ({
   baseUrl = '{{ .BaseURL }}',
   firstName = '{{ .FirstName }}',
   lastName = '{{ .LastName }}',
-  linkName = '{{ .LinkName }}',
-  linkURL = '{{ .LinkURL }}',
+  trialExpirationDate = '{{ .TrialExpirationDate }}',
+  trialExpirationWindow = '{{ .TrialExpirationWindow }}',
   supportEmail = '{{ .SupportEmail }}',
-}: PlaidDisconnectedProps) => {
-  const previewText = 'One of your linked accounts has been disconnected';
+}: TrialAboutToExpireProps) => {
+  const previewText = 'Your monetr trial is about to expire';
   return (
     <EmailLayout previewText={previewText}>
       <EmailLogo baseUrl={ baseUrl } />
       <Heading className='text-black text-2xl font-normal text-center p-0 my-8 mx-0'>
-        One of your linked accounts has been disconnected
+        Your trial for <strong>monetr</strong> is about to expire
       </Heading>
       <Text className='text-black text-sm leading-6'>
         Hello {firstName},
       </Text>
       <Text className='text-black text-sm leading-6'>
-        Your <strong>{linkName}</strong> account connected via Plaid needs to be reauthenticated. This account
-        will not receive automatic updates until the link has been updated.
+        We are just letting you know that your trial is about to expire. Don't worry, if you don't want to continue
+        using monetr then no action is required on your part. If you would like to continue using monetr though, you
+        will need to setup a subscription the next time you login.
       </Text>
-      <Section className='text-center mt-9 mb-9'>
-        <Button
-          className='bg-purple-500 rounded-lg text-white text-sm font-semibold no-underline text-center'
-          href={linkURL}
-        >
-          <Text className='text-sm text-white m-2'>
-            Reconnect {linkName}
-          </Text>
-        </Button>
-      </Section>
+      <Text className='text-black text-sm leading-6'>
+        Your trial will expire in about <strong>{ trialExpirationWindow }</strong> on <strong>{ trialExpirationDate }</strong>.
+      </Text>
+      <Text className='text-black text-sm leading-6'>
+        Thank you for trying out monetr!
+      </Text>
       <Hr className='border border-solid border-gray-200 my-6 mx-0 w-full' />
       <Text className='text-gray-500 text-xs leading-6'>
         This message was intended for{' '}
@@ -69,15 +64,15 @@ export const PlaidDisconnected = ({
   );
 };
 
-PlaidDisconnected.PreviewProps = {
+TrialAboutToExpire.PreviewProps = {
   baseUrl: 'https://my.monetr.dev',
   firstName: 'Elliot',
   lastName: 'Courant',
-  linkName: 'Navy Federal Credit Union',
-  linkURL: 'https://my.monetr.dev/bank_accounts/bac_123abc',
+  trialExpirationDate: 'Monday October 1st, 2024',
+  trialExpirationWindow: '3 days',
   supportEmail: 'support@monetr.local',
-} as PlaidDisconnectedProps;
+} as TrialAboutToExpireProps;
 
-export default PlaidDisconnected;
+export default TrialAboutToExpire;
